+++ opencore_legacy_patcher/constants.py
@@ -14,7 +14,7 @@ class Constants:
     def __init__(self) -> None:
         # Patcher Versioning
         self.patcher_version:                 str = "2.5.0"  # OpenCore-Legacy-Patcher
-        self.patcher_support_pkg_version:     str = "1.9.5"  # PatcherSupportPkg
+        self.patcher_support_pkg_version:     str = "1.9.6"  # PatcherSupportPkg
         self.copyright_date:                  str = "Copyright © 2020-2025 Dortania"
         self.patcher_name:                    str = "OpenCore Legacy Patcher"